refactor transfer logs method to reusable service trait
